fix(cli): point 3 stale consumers at the lifted src/cli command-registry (Move 1 follow-up)
buildFullCommandTree + the load* consts moved to src/cli/command-registry.ts,
but 3 importers still referenced the old lib/startup path and broke at runtime
(gen-command-index.test.ts, artifacts.test.ts, index.bench.ts). Re-exporting from
lib/startup was rejected: it would route lib->cli->commands->lib and reintroduce
the cycle this move breaks. Consumers now import from the new path. RUSH-2708.
